pyproject.toml
[project]
name = "devsync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.2",
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "django>=5.2",
    "django-cleanup>=9.0.0",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-redis>=5.4.0",
    "djangorestframework>=3.16.0",
    "djoser>=2.3.1",
    "drf-nested-routers>=0.94.1",
    "gunicorn>=23.0.0",
    "pillow>=11.2.1",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-json-logger>=3.3.0",
    "redis>=5.2.1",
    "uvicorn[standard]>=0.34.2",
]

[dependency-groups]
dev = []
